template/Data-Science/Prediction/Weather-Prediction/requirements.txt
requests
pandas
scikit-learn
xgboost
streamlit
python-dotenv